Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
dnspython==2.4.2
flask-limiter==3.5.0
google-auth==2.25.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
bcrypt==4.1.2
email-validator==2.1.0
gunicorn==21.2.0